Move did stuff to did module
